Loglikelihood adjustment for evir fits

S3 alogLik method to perform loglikelihood adjustment for fitted extreme value model objects returned from the functions gev, gpd and pot in the evir package. If x was returned from pot then the model will need to be re-fitted using pot_refit.

# S3 method for gev
alogLik(x, cluster = NULL, use_vcov = TRUE, ...)

# S3 method for gpd
alogLik(x, cluster = NULL, use_vcov = TRUE, ...)

# S3 method for potd
alogLik(x, cluster = NULL, use_vcov = TRUE, ...)

Arguments

x
A fitted model object with certain associated S3 methods. See Details.

cluster
A vector or factor indicating from which cluster the respective loglikelihood contributions from loglik originate. This must have the same length as the vector returned by the logLikVec method for an object like x. If cluster is not supplied (i.e. is NULL) then it is assumed that each observation forms its own cluster. See Details.

use_vcov
A logical scalar. Should we use the vcov S3 method for x (if this exists) to estimate the Hessian of the independence loglikelihood to be passed as the argument H to adjust_loglik? Otherwise, H is estimated inside adjust_loglik using optimHess.

...
Further arguments to be passed to the functions in the sandwich package meat (if cluster = NULL), or meatCL (if cluster is not NULL).

Value

An object inheriting from class "chandwich". See adjust_loglik. class(x) is a vector of length 5. The first 3 components are c("lax", "chandwich", "evir"). The remaining 2 components depend on the model that was fitted. If gev was used then these components are c("gev", "stat"). If gpd was used then these components are c("gpd", "stat"). If pot_refit was used then these components are c("potd", "stat").

Details

See alogLik for details.

If pot was used then x does not contain the raw data that alogLik needs. The model will need to be re-fitted using pot_refit and the user will be prompted to do this by an error message produced by alogLik.

References

Chandler, R. E. and Bate, S. (2007). Inference for clustered data using the independence loglikelihood. Biometrika, 94(1), 167-183. http://doi.org/10.1093/biomet/asm015

Zeleis (2006) Object-Oriented Computation and Sandwich Estimators. Journal of Statistical Software, 16, 1-16. http://doi.org/10.18637/jss.v016.i09

See also

Examples

# We need the evir package
got_evir <- requireNamespace("evir", quietly = TRUE)
if (got_evir) {
  library(evir)
  # An example from the evir::gev documentation
  data(bmw)
  out <- gev(bmw, "month")
  adj_out <- alogLik(out)
  summary(adj_out)

  # An example from the evir::gpd documentation
  data(danish)
  out <- gpd(danish, 10)
  adj_out <- alogLik(out)
  summary(adj_out)

  # An example from the evir::pot documentation
  # We use lax::pot_refit() to return the input data
  out <- pot_refit(danish, 10)
  adj_out <- alogLik(out)
  summary(adj_out)
}