Loglikelihood adjustment for fExtremes fits

S3 alogLik method to perform loglikelihood adjustment for fitted extreme value model objects returned from the functions gevFit, gumbelFit and gpdFit in the fExtremes package. The model must have been fitted using maximum likelihood estimation.

# S3 method for fGEVFIT
alogLik(x, cluster = NULL, use_vcov = TRUE, ...)

# S3 method for fGPDFIT
alogLik(x, cluster = NULL, use_vcov = TRUE, ...)

Arguments

x
A fitted model object with certain associated S3 methods. See Details.

cluster
A vector or factor indicating from which cluster the respective loglikelihood contributions from loglik originate. This must have the same length as the vector returned by the logLikVec method for an object like x. If cluster is not supplied (i.e. is NULL) then it is assumed that each observation forms its own cluster. See Details.

use_vcov
A logical scalar. Should we use the vcov S3 method for x (if this exists) to estimate the Hessian of the independence loglikelihood to be passed as the argument H to adjust_loglik? Otherwise, H is estimated inside adjust_loglik using optimHess.

...
Further arguments to be passed to the functions in the sandwich package meat (if cluster = NULL), or meatCL (if cluster is not NULL).

Value

An object inheriting from class "chandwich". See adjust_loglik. class(x) is a vector of length 5. The first 3 components are c("lax", "chandwich", "fExtremes"). The remaining 2 components depend on the model that was fitted. If gevFit or gumbelFit was used then these components are c("gev", "stat"). If gpdFit was used then these components are c("gpd", "stat").

Details

See alogLik for details.

References

Chandler, R. E. and Bate, S. (2007). Inference for clustered data using the independence loglikelihood. Biometrika, 94(1), 167-183. http://doi.org/10.1093/biomet/asm015

Zeleis (2006) Object-Oriented Computation and Sandwich Estimators. Journal of Statistical Software, 16, 1-16. http://doi.org/10.18637/jss.v016.i09

See also

Examples

# We need the fExtremes package
got_fExtremes <- requireNamespace("fExtremes", quietly = TRUE)
if (got_fExtremes) {
  library(fExtremes)

  # GEV
  # An example from the fExtremes::gevFit documentation
  set.seed(4082019)
  x <- gevSim(model = list(xi=0.25, mu=0, beta=1), n = 1000)
  # Fit GEV distribution by maximum likelihood estimation
  fit <- gevFit(x)
  adj_fit <- alogLik(fit)
  summary(adj_fit)

  # GP
  # An example from the fExtremes::gpdFit documentation
  # Simulate GP data
  x <- gpdSim(model = list(xi = 0.25, mu = 0, beta = 1), n = 1000)
  # Fit GP distribution by maximum likelihood estimation
  fit <- gpdFit(x, u = min(x))
  adj_fit <- alogLik(fit)
  summary(adj_fit)
}

#> Loading required package: timeDate

#> Loading required package: timeSeries

#> Loading required package: fBasics

#> Loading required package: fGarch

#> 
#> Attaching package: 'fExtremes'

#> The following objects are masked from 'package:evir':
#> 
#>     dgev, dgpd, pgev, pgpd, qgev, qgpd, rgev, rgpd

#> The following objects are masked from 'package:texmex':
#> 
#>     dgev, dgpd, pgev, pgpd, qgev, qgpd, rgev, rgpd

#> The following objects are masked from 'package:evd':
#> 
#>     dgev, dgpd, pgev, pgpd, qgev, qgpd, rgev, rgpd

#>         MLE      SE adj. SE
#> xi   0.2514 0.03844 0.04026
#> beta 0.9989 0.04909 0.04852